* add Lua 5.1 compatibility * Makefile - port to MSVC++2005 compiler and luabinaries * Makefile - add -DLUA_USE_APICHECK in debug build * fix compiler warnings (msvc2005/mingw/gcc3+4) * support compilation under linux/winelib (though doesn't really run)

+++ src/library/tStringBuffer.cpp
@@ -24,7 +24,7 @@ tStringBuffer::~tStringBuffer()
 
 void tStringBuffer::copyToBuffer(char * source)
 {
-  long new_size = strlen(source) + 1;
+  size_t new_size = strlen(source) + 1;
 
   if(new_size > size)
   {
